Update clickhouse-rs to reduce memory usage and CPU consumption
Slow query log reads lots of rows (with --cluster and lots of nodes in
the cluster, you will get X*100 at most, and with cluster of 100 nodes
it will be 10000 rows), each will have HashMap for ProfileEvents, which
is very greedy in upstream clickhouse-rs, the problem was that the
capacity that was used to create HashMap was incorrect - it was not the
number of elements, but instead, offset from the beginning.

This has been fixed in [1], but for now I will update the crate source
to my fork.

  [1]: https://github.com/suharev7/clickhouse-rs/pull/19