Catch some query errors (by reworking error handling of queue messages)
Previously the chdig crashes on EXPLAIN PLAN for INSERT queries (since
INSERT does not support EXPLAIN PLAN), and this leads to crashing of the
worker thread (that handles queue messages for running queries in
ClickHouse and triggering UI).

Reworking of the error handling in the worker makes the code cleaner
(and even reduces amount of code) and automatic error handling.