Better error handling of failed UI message passing
It is possible for cursive's cb_sink.send() to fail, in case of it had
been destroyed already, but we should not fail in this case, the error
should be ignored, like one place in the code already did.

But only one place had been modified, while there is tons of them.

So make process_event() to return the error, and the caller will handle
the error (or ignore it).