Speedup query for extracting slow queries (makes query 10x faster)
The columns are too heavy, it is better to use PREWHERE, since by
default query_id IN did not moved into PREWHERE.

And now on some obscure cluster it runs 1 second instead of 10.

But note, that even though the query runs 1 second, processing the rows
by chdig takes significant amount of time:
- debug: 10sec (10x slower)
- release: 6-7sec (6-7x slower)

And this, should be optimized!