Implement simple processes/queries filtering (by users and client info)
I wanted to do this long time ago, but the problem was the interface
part. But I have a "briliant" idea on how this could be done, instead of
figuring out what the interface should be I will just use OR for all
interesting columns and done with it!

Here is a list of columns:
- client_hostname
- os_user
- user
- initial_user
- client_name