Use termion buffered backend to avoid dependency from cursive
There are some issues with depeendency from the ncurses library, it does
not allows to compile chdig in a static binary (this could be done, but
requires some work), and this leads to a requirement for minimal glibc
version, recently I tried to run chdig on debian buster, and it fails:

    $ chdig
    chdig: /lib/x86_64-linux-gnu/libm.so.6: version `GLIBC_2.29' not found (required by chdig)
    chdig: /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.29' not found (required by chdig)

    # apt-cache policy libc6
    libc6:
      Installed: 2.28-10+deb10u2
      Candidate: 2.28-10+deb10u2
      Version table:
     *** 2.28-10+deb10u2 500
            500 http://security.debian.org/debian-security buster/updates/main amd64 Packages
            100 /var/lib/dpkg/status
         2.28-10+deb10u1 500
            500 http://deb.debian.org/debian buster/main amd64 Packages

Last time I've tried other backends, they all experience flickering
issues [1], but, cursive_buffered_backend [2] - solves the problem!

  [1]: https://github.com/gyscos/Cursive/issues/142
  [2]: https://github.com/agavrilov/cursive_buffered_backend

Also I have to remove --mouse/--no-mouse, since for this the cursive
should be patched.