Change the way focus works in tables - allow to stick to the top
It is not always useful to always select some item (i.e. first by
sorting order, by the way it is not even how it works natively, I need
to set_selected_item() call for this before).

Consider csysdig, by default when you did not press anything it stick to
the first item by sort order, and only if you select something (Up/Down)
then it will stick to this item in the table.

And this behaviour is better, since if you did not focus any item you
want to look at the items sorted by the sort order.

And I also made the Home button to cancel this selection.