Add ability to edit and execute query
OLAP quieries can be pretty big, and may not fit into one screen, and
some of them even exceeds default pretty print limits, so it is not
always easy to copy-paste them for experimenting.

So this patch implements spawning editor with the query and after
executing it, that way you can experiment with various settings and
other things (note, that all SETTINGS will be automatically appended
into the query).