Embedded flameshow (fixes #6)
flameshow is a written in python library, that right now is compiled
into a single binary using PyOxidize.

However it will be "cool" to embed it into chdig binary, so that we can
ship one single binary -- chdig.

But PyOxidize does not provide support for compiling libraries, so some
trickery is required:
- we need python distribution with resources that is defined in Starlack
  (pyoxidize resources)
- we need python distribution (pyoxidize generate-python-embedding-artifacts)
- pyo3 configure incorrect config that has libm, this should be replaced
- PyOxidize configure config that uses incorrect packed-resources (not
  the one that has our resources from Starlack, but a simple one with
  python only)
- and the cherry on the top is that you need to have pyo3 config
  *before* running main cargo build -- that's why there is a separate
  `cargo build -p flameshow`

See also comments in flameshow/src/lib.rs and flameshow/build.rs (but in
a nutshell Python in Rust is not that stable, maybe I will have to
revert this patch eventually, but this was fun anyway!)