Support skip_verify in client config to skip strict TLS verification
Many a times like in case of self-signed certificate or sometimes with
letsencrypt certs, TLS verification fails with unknown issuer error. In
those cases, it is useful to enable skip_verify

Note, that currently client does not support this, instead you need to
use something more complex, like:

    <config>
        <openSSL>
            <client>
                <loadDefaultCAFile>true</loadDefaultCAFile>
                <cacheSessions>true</cacheSessions>
                <disableProtocols>sslv2,sslv3</disableProtocols>
                <preferServerCiphers>true</preferServerCiphers>
                <invalidCertificateHandler>
                    <name>AcceptCertificateHandler</name>  <!-- This line is analog of skip_verify -->
                </invalidCertificateHandler>
            </client>
        </openSSL>
    <config>

But it is a good thing to add a alias for it I guess (not to export all
the Poco internals).